--- app/src/main/java/com/example/algorithmstudy/pattern/creation/abstract_factory/Infantry.kt
@@ -1,6 +1,0 @@
-package com.example.algorithmstudy.pattern.creation.abstract_factory
-
-import com.example.algorithmstudy.pattern.creation.abstract_factory.Unit
-
-interface Infantry : Unit {
-}
