--- app/src/main/java/com/example/algorithmstudy/pattern/creation/abstract_factory/InfantryUnits.kt
@@ -1,6 +1,0 @@
-package com.example.algorithmstudy.pattern.creation.abstract_factory
-
-enum class InfantryUnits {
-    RIFLEMAN,
-    ROCKET_SOLDIER
-}
