--- app/src/main/java/com/example/algorithmstudy/pattern/creation/abstract_factory/Vehicle.kt
@@ -1,4 +1,0 @@
-package com.example.algorithmstudy.pattern.creation.abstract_factory
-
-interface Vehicle : Unit {
-}
