--- app/src/main/java/com/example/algorithmstudy/pattern/creation/abstract_factory/units/RocketSoldier.kt
@@ -1,6 +1,0 @@
-package com.example.algorithmstudy.pattern.creation.abstract_factory.units
-
-import com.example.algorithmstudy.pattern.creation.abstract_factory.Infantry
-
-class RocketSoldier: Infantry {
-}
