+++ app/src/main/java/com/example/algorithmstudy/pattern/creation/abstractfactory/Building.kt
@@ -1,4 +1,4 @@
-package com.example.algorithmstudy.pattern.creation.abstract_factory
+package com.example.algorithmstudy.pattern.creation.abstractfactory
 
 interface Building<in UnitType, out ProducedUnit>
         where UnitType : Enum<*>, ProducedUnit : Unit {
