+++ app/src/main/java/com/example/algorithmstudy/pattern/creation/abstractfactory/InfantryUnits.kt
@@ -1,0 +1,6 @@
+package com.example.algorithmstudy.pattern.creation.abstractfactory
+
+enum class InfantryUnits {
+    RIFLEMAN,
+    ROCKET_SOLDIER
+}
