+++ app/src/main/java/com/example/algorithmstudy/pattern/creation/abstractfactory/Vehicle.kt
@@ -1,0 +1,4 @@
+package com.example.algorithmstudy.pattern.creation.abstractfactory
+
+interface Vehicle : Unit {
+}
