+++ app/src/main/java/com/example/algorithmstudy/pattern/creation/abstractfactory/units/Rifleman.kt
@@ -1,0 +1,6 @@
+package com.example.algorithmstudy.pattern.creation.abstractfactory.units
+
+import com.example.algorithmstudy.pattern.creation.abstractfactory.Infantry
+
+class Rifleman: Infantry {
+}
