+++ app/src/test/java/com/example/algorithmstudy/design_pattern/PrototypeTest.kt
@@ -3,6 +3,11 @@ package com.example.algorithmstudy.design_pattern
 import com.example.algorithmstudy.pattern.creation.prototype.MyComputer
 import org.junit.Test
 
+/**
+ *  Prototype Pattern
+ *
+ *  - 이미 생성된 객체의 일부를 복사해 사용하고자 하는 경우 쓰면 유용한 패턴
+ */
 class PrototypeTest {
 
     @Test
@@ -16,6 +21,5 @@ class PrototypeTest {
 
         println(myOriginalPc)
         println(myOriginalPc.copy(graphicCard = "Rygen"))
-
     }
 }
